fix: Corrected SchemaFactoryMakeCommand.php to work with Laravel 11 (#15)
* fix: Corrected SchemaFactoryMakeCommand.php to work with Laravel 11

* fix: Update SchemaFactoryMakeCommand.php to work with Laravel 11

Changed way to get $default and $notNull, to be more widely compatible

* fix: Update SchemaFactoryMakeCommand.php to work with Laravel 11

Changed way to get $default and $notNull, to be more widely compatible and tested working

* fix: Update SchemaFactoryMakeCommand to improve new code readability

Refactored the code to use the schema builder to retrieve the columns of the table and iterate over each column to find the specific one we're interested in. This change ensures compatibility with Laravel 11 and provides a more widely compatible and tested solution.